separated classes and cli functions
